refactor(map): separate map related logic
